Remove duplicate files and fix problems with tests
